Reorganize consolidated dispatcher code, no behavior change
